+++ Assets/LoadData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 78846e842a6cc0b4ba3d0e7b971365a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
