+++ Assets/_UDVT/Scripts/Runtime/MenuScripts/CurrentParams.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-public class CurrentVisType : MonoBehaviour
+public class CurrentParams : MonoBehaviour
 {
     public static VisType currentVisType = VisType.Scatterplot; // default
+    public static Dictionary<string, double[]> loadedData = null; // default
 }
