+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,9 @@ EditorBuildSettings:
     path: Assets/_UDVT/Scenes/Menu.unity
     guid: 2a9782eaa5395964098899bcfa078ad6
   - enabled: 1
+    path: Assets/_UDVT/Scenes/LoadData.unity
+    guid: 99504c0a53b9e3c49bad290bfa6bb24d
+  - enabled: 1
     path: Assets/_UDVT/Scenes/MainScene.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
   m_configObjects: {}
