+++ UserSettings/EditorUserSettings.asset
@@ -9,6 +9,12 @@ EditorUserSettings:
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-1:
+      value: 0152515606565d080f5c5e2745215a4417164d78792c7e657c7e4d63bab06739
+      flags: 0
+    RecentlyUsedSceneGuid-2:
+      value: 5a0801565156595a585d0e7b15705c444f15192e7e7176322b284b31e0b2603c
+      flags: 0
+    RecentlyUsedSceneGuid-3:
       value: 51500d515d070c5a0c5b5f7b457a0944464e4072757124332b284d64bae1306e
       flags: 0
     vcSharedLogLevel:
